Code reorganization part #1
o Moved functions and declarations from titlekey.c / titlekey.h into
padgen.c / padgen.h
o Moved "helper functions to the top, feature functions to the bottom
o Made a separate, clutter free header file for feature functions
o Moved the #defines to the header file
o Renamed MAXENTRIES to MAX_ENTRIES for consistency
