+++ source/decryptor/padgen.h
@@ -2,7 +2,10 @@
 
 #include "common.h"
 
-#define MAXENTRIES 1024
+#define BUFFER_ADDRESS	((u8*) 0x21000000)
+#define BUFFER_MAX_SIZE	(1 * 1024 * 1024)
+
+#define MAX_ENTRIES 1024
 
 typedef struct {
     u8   CTR[16];
@@ -12,7 +15,7 @@ typedef struct {
 
 typedef struct {
     u32 n_entries;
-    SdInfoEntry entries[MAXENTRIES];
+    SdInfoEntry entries[MAX_ENTRIES];
 } __attribute__((packed, aligned(16))) SdInfo;
 
 
@@ -32,8 +35,9 @@ typedef struct {
     u32 ncch_info_version;
     u32 n_entries;
     u8  reserved[4];
-    NcchInfoEntry entries[MAXENTRIES];
+    NcchInfoEntry entries[MAX_ENTRIES];
 } __attribute__((packed, aligned(16))) NcchInfo;
+
 
 typedef struct {
     u64 titleId;
@@ -44,7 +48,7 @@ typedef struct {
 typedef struct {
     u32 n_entries;
     u8 padding[12];
-    SeedInfoEntry entries[MAXENTRIES];
+    SeedInfoEntry entries[MAX_ENTRIES];
 } __attribute__((packed)) SeedInfo;
 
 
@@ -58,8 +62,15 @@ typedef struct {
 } __attribute__((packed, aligned(16))) PadInfo;
 
 
-u32 NcchPadgen(void);
-u32 SdPadgen(void);
-u32 NandPadgen(void);
+typedef struct {
+	u32 commonKeyIndex;
+	u8  reserved[4];
+	u8  titleId[8];
+	u8  encryptedTitleKey[16];
+} __attribute__((packed)) TitleKeyEntry;
 
-u32 CreatePad(PadInfo *info);
+typedef struct {
+	u32 n_entries;
+	u8  reserved[12];
+	TitleKeyEntry entries[MAX_ENTRIES];
+} __attribute__((packed, aligned(16))) EncKeysInfo;
